REFACTOR: refactored codebase into multiplatform structure
