+++ App/src/commonMain/graphql/TopAnimeAiring.graphql
