+++ App/src/commonMain/graphql/TopAnimeByRating.graphql
