+++ App/src/commonMain/graphql/TopAnimeBySeason.graphql
@@ -1,0 +1,22 @@
+query TopAnimeBySeason($page: Int, $perPage: Int, $season: MediaSeason, $seasonYear: Int) {
+    Page(page: $page, perPage: $perPage) {
+        media(sort: SCORE_DESC, type: ANIME, season: $season, seasonYear: $seasonYear) {
+            title {
+                english
+                native
+            }
+            id
+            status
+            episodes
+            bannerImage
+            averageScore
+            seasonYear
+            season
+            description
+            genres
+            coverImage {
+                large
+            }
+        }
+    }
+}
