+++ App/src/desktopMain/kotlin/dev/dubsky/aiko/components/button/ActionButton.kt
@@ -1,0 +1,32 @@
+package dev.dubsky.aiko.components.button
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import org.jetbrains.compose.ui.tooling.preview.Preview
+
+@Preview
+@Composable
+fun ActionButton(color: Color, onClick: () -> Unit) {
+    Box(
+        modifier = Modifier
+            .size(15.dp)
+            .clip(RoundedCornerShape(10.dp))
+            .background(color)
+            .clickable(
+                interactionSource = remember { MutableInteractionSource() },
+                indication = null
+            ) {
+                onClick()
+            }
+    )
+}
