+++ App/src/desktopMain/kotlin/dev/dubsky/aiko/data/Anime.kt
