+++ App/src/desktopMain/kotlin/dev/dubsky/aiko/screens/ComingScreen.kt
@@ -1,0 +1,12 @@
+package dev.dubsky.aiko.screens
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ComingScreen() {
+    MaterialTheme {
+        Text("Coming soon...")
+    }
+}
