+++ gradle.properties
@@ -1,4 +1,6 @@
-org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
+#Kotlin
 kotlin.code.style=official
-kotlin.version=2.0.21
-compose.version=1.7.0
+kotlin.daemon.jvmargs=-Xmx2048M
+
+#Gradle
+org.gradle.jvmargs=-Xmx2048M -Dfile.encoding=UTF-8
