--- src/main/kotlin/dev/dubsky/aiko/components/bar/ControlBar.kt
@@ -1,69 +1,0 @@
-package dev.dubsky.aiko.components.bar
-
-import androidx.compose.desktop.ui.tooling.preview.Preview
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.WindowState
-import dev.dubsky.aiko.components.button.ActionButton
-
-@Preview
-@Composable
-fun ControlBar(
-    windowState: WindowState,
-    onMinimizeClick: () -> Unit,
-    onMaximizeClick: () -> Unit,
-    onCloseClick: () -> Unit
-) {
-    Box(
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(40.dp)
-                .background(Color(0xFF0F141A))
-        ) {
-            Text(
-                text = "v0.1.0",
-                color = Color.White.copy(alpha = 0.6f),
-                fontSize = 12.sp,
-                modifier = Modifier
-                    .align(Alignment.CenterStart)
-                    .padding(start = 16.dp)
-            )
-
-            Text(
-                text = "Aiko",
-                color = Color.White,
-                fontSize = 16.sp,
-                fontWeight = FontWeight.Medium,
-                modifier = Modifier.align(Alignment.Center)
-            )
-
-            Row(
-                modifier = Modifier
-                    .align(Alignment.CenterEnd)
-                    .padding(end = 8.dp),
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-
-                ActionButton(Color.Green, onClick = onMinimizeClick)
-                ActionButton(Color.Yellow, onClick = onMaximizeClick)
-                ActionButton(Color.Red, onClick = onCloseClick)
-            }
-        }
-    }
-}
