--- src/main/kotlin/dev/dubsky/aiko/screens/ComingScreen.kt
@@ -1,14 +1,0 @@
-package dev.dubsky.aiko.screens
-
-import androidx.compose.desktop.ui.tooling.preview.Preview
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-
-@Preview
-@Composable
-fun ComingScreen() {
-    MaterialTheme {
-        Text("Coming soon...")
-    }
-}
